+++ src/services/SendoutsService.php
@@ -260,6 +260,8 @@ class SendoutsService extends Component
                     Craft::$app->getQueue()->push(new SendoutJob([
                         'sendoutId' => $sendout->id,
                         'title' => $sendout->title,
+                        'unlimitedMemoryLimit' => Campaign::$plugin->getSettings()->unlimitedMemoryLimit,
+                        'unlimitedTimeLimit' => Campaign::$plugin->getSettings()->unlimitedTimeLimit,
                     ]));
 
                     $sendout->sendStatus = SendoutElement::STATUS_QUEUED;